fix: comprehensive accessibility audit and improvements
- Add skip-to-main-content link
- Add aria-hidden on decorative elements
- Add focus-visible ring styles on all interactive elements
- Fix form label/input associations with htmlFor/id
- Add role=tablist and aria-selected on AI widget tabs
- Add scope attributes on comparison table headers
- Add sr-only text for icon-only check/minus in tables
- Add metadata descriptions on pages missing them
- Extract OptInForm to separate client component
- Add aria-labels on buttons without visible text
