+++ components/sections/CTASection.jsx
@@ -115,7 +115,7 @@ export function CTASection() {
             {/* Primary: white bg */}
             <Link
               href="/contact"
-              className="inline-flex items-center gap-2 no-underline text-sm font-semibold rounded-[11px] px-6 py-3.5 transition-all duration-200 hover:-translate-y-0.5 hover:shadow-xl"
+              className="inline-flex items-center gap-2 no-underline text-sm font-semibold rounded-[11px] px-6 py-3.5 transition-all duration-200 hover:-translate-y-0.5 hover:shadow-xl focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white/60 focus-visible:ring-offset-2 focus-visible:ring-offset-primary"
               style={{
                 fontFamily: 'var(--font-outfit), Outfit, sans-serif',
                 background: '#FFFFFF',
@@ -130,7 +130,7 @@ export function CTASection() {
             {/* Secondary: frosted */}
             <Link
               href="/products/receptionist"
-              className="inline-flex items-center gap-2 no-underline text-sm font-semibold text-white rounded-[11px] px-6 py-3.5 transition-all duration-200 hover:-translate-y-0.5"
+              className="inline-flex items-center gap-2 no-underline text-sm font-semibold text-white rounded-[11px] px-6 py-3.5 transition-all duration-200 hover:-translate-y-0.5 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-white/60 focus-visible:ring-offset-2 focus-visible:ring-offset-primary"
               style={{
                 fontFamily: 'var(--font-outfit), Outfit, sans-serif',
                 background: 'rgba(255,255,255,0.12)',
